Xong design thêm, cập nhật phim, loại phim, rạp. Xong model, file seed (thiếu của bảng vé). Cập nhật migrate

+++ database/migrations/2020_11_26_161406_create_nhan_viens_table.php
@@ -29,7 +29,7 @@ class CreateNhanViensTable extends Migration
             $table->bigInteger('MaNQL')->nullable()->unsigned(); //khóa ngoại
             $table->string('TenTK', 20);
             $table->string('Pass', 20);
-            $table->integer('TrangThai')->default(0);
+            $table->tinyInteger('TrangThai')->default(0);
             $table->timestamp('ThoiGianTao')->useCurrent();
             $table->timestamp('ThoiGianCapNhatCuoi')->useCurrent();
         });
